feat(cohorts): update and delete

+++ src/services/programsService.js
@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this */
 import models from '../database/models';
 
-const { program, cohort } = models;
+const { program, cohort, user, rating } = models;
 
 class programsService {
   static async create(programData) {
@@ -34,7 +34,18 @@ class programsService {
   }
 
   static async removeOne(id) {
-    return program.destroy({ where: { id } });
+    const found = await program.findOne({
+      where: { id },
+      include: [{ model: user, as: 'users' }],
+    });
+    if (found.users.length) {
+      return { error: 'Some users are using this program!' };
+    }
+    const rate = await rating.findOne({ where: { program: found.id } });
+    if (rate) return { error: 'Please remove ratings associated to program' };
+    const deleted = await program.destroy({ where: { id } });
+    if (!deleted) return { error: 'Unable to delete program!' };
+    return { deleted: found };
   }
 
   static async exists(id) {
